adding .DS_Store to gitignore, moving hooks & lib & utils folder to within src folder
