+++ src/pages/_app.tsx
@@ -1,6 +1,6 @@
 import type { AppProps } from 'next/app';
 import { ApolloProvider } from '@apollo/client';
-import useApollo from '../../hooks/useApollo.ts';
+import useApollo from '../hooks/useApollo.ts';
 
 const App = ({ Component, pageProps }: AppProps) => {
   const client = useApollo(pageProps);
